Fixes an edge case with org reuse
The tree view controller works by querying ApexClass, and sorting the found classes by the doc block’s @group tag. When users re-use a sandbox or other org that pre-contains Apex classes that *do not* have a @group tag in the doc block, it can cause a Regex governor limit issue.

This addresses that by moving from a SOQL query, to a SOSL query looking for the @group tag in a doc block.